requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyPDF2==3.0.1
sentence-transformers==2.2.2
torch==2.1.2
faiss-cpu==1.7.4
numpy==1.24.4
huggingface-hub==0.20.3
transformers==4.37.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-Limiter==3.5.1
